Move WithHoverOrFocusProp
