+++ packages/lib-classifier/src/components/Classifier/components/TaskArea/components/Answer/components/AnswerButton/AnswerButton.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
 
-import WithHoverOrFocusProp from '../../../../../../WithHoverOrFocusProp'
+import WithHoverOrFocusProp from '../../../../../WithHoverOrFocusProp'
 
 const StyledButton = styled.button`
   border: 0;
