+++ packages/lib-classifier/src/components/Classifier/components/WithHoverOrFocusProp/WithHoverOrFocusProp.js
